glibc-tests: Fix failing tst-cond4 and tst-cond6 tests
tst-cond4 and tst-cond6 as imported from glibc rely on an undefined
behavior of process shared mutexes. In the original implementation 'mut1'
is locked by the parent thread before fork and it is unlocked by the child
thread after acquiring 'mut2'. This is done to ensure that the
pthread_cond_signal() call happens after the pthread_cond_wait() call.
However the child process should not be owning 'mut1' per latest POSIX
clarification[1]:

"Any locks held by any thread in the calling process that have been set to
be process-shared shall not be held by the child process. For locks held by
any thread in the calling process that have not been set to be
process-shared, any attempt by the child process to perform any operation
on the lock results in undefined behavior (regardless of whether the
calling process is single-threaded or multi-threaded)."

[1] http://austingroupbugs.net/view.php?id=1112

Fix these tests by swapping the logic between the parent and child threads.
In the new implementation:

  1. parent locks 'mut1' and calls fork()

  2. child blocks on acquiring 'mut1'

  3. parent locks 'mut2' and unlocks 'mut1'

  4. child locks 'mut1' and blocks on acquiring 'mut2'

  5. parent (timed) waits on 'cond' atomically releasing 'mut2'

  6. child can now acquire 'mut2', signal the conditional variable 'cond'
  and release 'mut2'

  7. parent wakes-up holding 'mut2', unlocks it and finishes execution

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>