glibc-tests: Reuse tst-cond4 implementation for tst-cond6
tst-cond6 implementation is virtually identical to tst-cond4 with the
exception that tst-cond6 is testing timed waits for conditional variables
vs. regular waits in tst-cond4.

Take advantage of this fact by including the tst-cond4 code in tst-cond6
with an additional 'TIMED' pre-processor define used to enable the
pi_cond_timedwait() code path. This is the same strategy already in use by
tst-cond20 and tst-cond21.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>