pi_cond: Eliminate the pending_wake counter
The 'pending_wake' counter is used to guard against spurious wake-ups
for a waiter which failed to enqueue with the kernel via a
futex_wait_requeue_pi() call (i.e. kernel returns EAGAIN) or for
normal wake-ups when all wake-ups have been consumed by other threads.

This adds a lot of complexity and currently causes the following tests to
fail:

  - tests/glibc-tests/tst-cond18
  - tests/glibc-tests/tst-cond21
  - tests/glibc-tests/tst-cond24

In addition to the added complexity and test failures this counter also
introduces a subtle priority inversion in the EAGAIN case: if multiple
waiters and a waker race with each other a lower priority waiter can grab
the private lock first and consume the wake-up signal that should have gone
to the highest priority waiter.

In the end it is actually simpler and safer to let the user
application handle the spurious wake-up rather than miss a wake-up.
This is the expected behavior per POSIX[1]:

  "When using condition variables there is always a Boolean predicate
  involving shared variables associated with each condition wait that
  is true if the thread should proceed. Spurious wakeups from the
  pthread_cond_timedwait() or pthread_cond_wait() functions may occur.
  Since the return from pthread_cond_timedwait() or
  pthread_cond_wait() does not imply anything about the value of this
  predicate, the predicate should be re-evaluated upon such return."

Simplify the implementation by eliminating the pending_wake counter.
Also enable the tests which are now passing.

[1] https://pubs.opengroup.org/onlinepubs/9699919799/functions/pthread_cond_wait.html

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>