pi_cond: Streamline pi_cond_timedwait and pi_cond_signal_common
Now that we have eliminated the complexity associated with the
pending_wait/wake counters we can make the code easier to read by
eliminating the do..while() loops in pi_cond_timedwait() and
pi_cond_signal_common() functions and using goto for retrying the EGAIN
error cases.

This reduces the indentation level and makes the code easier to read and
understand. No functional changes.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>