pi_cond: Remove private mutex and rely instead on user mutex
The librtpi API requires holding the user mutex when waiting or signaling a
conditional variable. Take advantage of this fact to eliminate the need for
an internal private mutex when accessing the condvar internals i.e. the
user mutex already gives us exclusive access.

Remove the internal private lock. This provides performance improvements
and also avoids a lot of error cases we would otherwise have to deal with
when the operations on the internal lock fail.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>