pi_cond: Add state tracking
The current implementation of pi_cond_init()/pi_cond_destroy() has no state
tracking. This is not usually a problem because all rtpi condvar internals
are guarded by the user mutex. The exceptions are pi_cond_wait() and
pi_cond_timedwait() which have to drop the user mutex before waiting. This
opens a race window where a concurrent thread can call pi_cond_destroy()
and destroy the condvar from under the waiting thread.

Add a .state field to the pi_cond_t which allows us to detect if a destroy
call has happened while waiting on the condvar.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>