glibc-tests, tst-cond20: Release mutex after signal/broadcast
Commit 6ff45b62ca0e ("glibc-tests: Fix races on condvar destroy")
erroneously moved the pi_cond_destroy() call into the same block of code
which does the 'cond' condvar signaling under mutex 'mut' protection.

This in effect blocks all waiters on the 'cond' from acquiring the mutex
and completing the wake-up until after pi_cond_destroy() is called. This
was not the original intent of this test.

Release the 'mut' which gives waiters a chance to race for it. Re-acquire
it before calling pi_cond_destroy() per librtpi API requirements.

As a side effect this makes the increased timeout introduced by commit
da49252f26c5 ("glibc-tests: Increase timeout for tst-cond20") unnecessary.
Remove it.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>